requirements.txt
delta-rest-client
pandas
pandas_ta
python-dotenv
requests
